% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_heading.R
\name{leer_encabezado}
\alias{leer_encabezado}
\title{Read multi row heading}
\usage{
leer_encabezado(df, inicio)
}
\arguments{
\item{df}{a tibble del cual se quiere leer el encabezado}

\item{inicio}{un vector que indica números de fila donde comienza el encabezado y los datos}
}
\value{
un vector con los nombres
}
\description{
\code{read_headings} returns a string vector with concatenated names of a multi row heading.
}
