% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leerplanilla_unavez.R
\name{leer_planilla}
\alias{leer_planilla}
\title{Leer planilla de priorización}
\usage{
leer_planilla(path)
}
\arguments{
\item{path}{}
}
\value{
Una lista de tibbles con las planillas para cada área
}
\description{
Leer planilla de priorización
}
