% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transferencias_fonacide}
\alias{transferencias_fonacide}
\title{Montos de Transferencias de FONACIDE}
\format{A data frame with 53940 rows and 10 variables:}
\usage{
transferencias_fonacide
}
\description{
Datos sobre las transferencias de FONACIDE a gobiernos descentralizados 
disponibilizados por el Ministerio de Hacienda en
\url{https://datos.hacienda.gov.py/data/fonacide-transferencia/descargas}.
}
\keyword{datasets}
