% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_encabezado.R
\encoding{UTF-8}
\name{leer_encabezado}
\alias{leer_encabezado}
\title{Lee el encabezado de varias filas}
\usage{
leer_encabezado(df, inicio)
}
\arguments{
\item{df}{un tibble del cual leer los encabezados}

\item{inicio}{un numeric vector de tamaño 2 que contiene la posición en `df` donde
empiezan el encabezado (\code{inicio[[1]]}) y los datos (\code{inicio[[1]]}).}
}
\value{
un string vector con los nombres de las columnas.
}
\description{
\code{leer_encabezado()} devuleve un string vector con los nombres de las columnas.
Concatena los nombres con las filas inferiores en caso que el encabezado tenga varias
filas. Se utiliza principalmente para leer los encabezados de las planillas de 
priorización.
}
