% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_encabezado.R
\name{leer_encabezado}
\alias{leer_encabezado}
\title{Read multi row heading}
\usage{
leer_encabezado(df, inicio)
}
\arguments{
\item{df}{a tibble to read the headings from}

\item{inicio}{a numeric vector length 2 that contains the row number where the
heading (position 1) and the data (position 2) start.}
}
\value{
a string vector that contains the concatenated columns names in the heading.
}
\description{
\code{read_headings} returns a string vector with concatenated names of a
multi row heading.
}
