% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_planilla.R
\encoding{UTF-8}
\name{leer_planilla}
\alias{leer_planilla}
\title{Leer planilla de priorización}
\usage{
leer_planilla(path, inicio_encabezado = "Prioridad", inicio_datos = "^1$")
}
\arguments{
\item{path}{ruta al archivo a leer}

\item{inicio_encabezado}{expresión regular para detectar el inicio del encabezado}

\item{inicio_datos}{expresión regular para detectar el inicio de los datos}
}
\value{
Una lista de tibbles con las planillas para cada área
}
\description{
Lee la planilla de priorización en formato excel y devuelve una
lista con tibbles para cada área de FONACIDE.
}
